Forms parcel order

+++ db/migrate/20200303134721_add_fields_to_user.rb
@@ -4,7 +4,7 @@ class AddFieldsToUser < ActiveRecord::Migration[5.2]
     add_column :users, :phone_number, :string
     add_column :users, :admin, :boolean
     add_column :users, :location, :string
-    add_column :users, :latitude, :decimal
-    add_column :users, :longitude, :decimal
+    add_column :users, :latitude, :float
+    add_column :users, :longitude, :float
   end
 end
